requirements.txt
astroid==2.5.6
autopep8==1.5.7
colorama==0.4.4
imutils==0.5.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
numpy==1.20.3
opencv-contrib-python==4.5.2.52
Pillow==8.2.0
pycodestyle==2.7.0
pylint==2.8.3
toml==0.10.2
wrapt==1.12.1
